Added the Health Script
This may or may not be used, but I'm bored in Calculus so I made it.
I feel like it gets rid of a lot of rewritten code.
If we add it to enemies and the player (and environment) then alert the
gamemaster that the object has died it would make our lives easier
later.